Got the page to show up
Added a route to the page
