+++ src/css/registration/SignUp.scss
@@ -2,11 +2,10 @@
   outline: none;
 }
 
-body {
-  font-family: Arial, Helvetica, sans-serif;
-  background-color: #0A090C;
-  color: white;
-  height: 100vh;
+#title {
+  padding-top: 2em;
+  padding-bottom: 1em;
+  text-align: center;
 }
 
 form {
@@ -14,6 +13,7 @@ form {
   align-items: center;
   justify-content: center;
   flex-direction: column;
+  padding-left: 4em;
 }
 
 label {
@@ -35,8 +35,9 @@ input:focus {
 }
 
 h6 {
-  color: grey;
+  color: white;
   font-weight: 100;
+  font-style: oblique;
   padding-bottom: 5px;
 }
 
@@ -70,63 +71,22 @@ button {
   font-size: 1.2em;
   transition: 1s;
   position: relative;
-  left: 34%;
+  margin-left: 7.5em;
 }
 
-button #comment {
-  display: none;
-}
 
 button:hover {
   background-color: #2E284E;
-}
-
-button:hover #comment {
-  display: none;
-}
-
-button:hover #newComment {
-  display: inline;
 }
 
 button:disabled {
   background-color: #454545;
 }
 
-h1 {
-  padding-top: 10px;
-  padding-bottom: 2%;
-  text-align: center;
-}
-
 #failedToRegister {
   color: red;
 }
 
-@media screen and (max-width: 321px) {
-  input {
-    width: 300px;
-    padding-left: 15px;
-    margin-left: 10px;
-  }
-
-  button {
-    left: 25%;
-  }
-
-  label {
-    margin-left: 15px;
-  }
-
-  p {
-    margin-left: 15px;
-  }
-
-  h6 {
-    padding-bottom: 5px;
-    margin-left: 15px;
-  }
-}
 
 .warning {
   padding-top: 15px;
